% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_data_by_analysis_field.r
\name{arrange_data_by_analysis_field}
\alias{arrange_data_by_analysis_field}
\title{Merges dart and meta data objects into a single object}
\usage{
arrange_data_by_analysis_field(dart_data, field, basedir, species, dataset)
}
\arguments{
\item{dart_data}{-- a merged dart-meta data object [Required]}

\item{field}{-- vector data field names for splitting [Required]}

\item{basedir}{-- name of the base directory [Required]}

\item{species}{-- species name [Required]}

\item{dataset}{-- dataset identifier [Required]}
}
\value{
object   -- a list of file names for the data objects that are produced
}
\description{
Input is a dart data object with altcount genotype encoding
(0=hom ref, 1=het, 2=hom alt).
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
