% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_allele_seqs.r
\name{cluster_allele_seqs}
\alias{cluster_allele_seqs}
\title{identifies clusters of loci base on a reference allele sequence}
\usage{
cluster_allele_seqs(
  base_fa,
  base_cl,
  species,
  dataset,
  cutoff = cutoff,
  cdhdir
)
}
\arguments{
\item{species}{-- name of the species in GenuSpec format [required]}

\item{dataset}{-- arbitrary name for the dataset [required]}

\item{cutoff}{-- cutoff for clustering algorithm [required]}

\item{basedir}{-- name of base directory [required]}

\item{trimmed}{-- use trimmed sequences (default = FALSE)}
}
\description{
identifies clusters of loci base on a reference allele sequence
}
\examples{
write.dart.data(dart_data, RandRbase, "FicuRubi", "DFi16-2154")
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
