% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_sequences_cdhit.r
\name{cluster_sequences_cdhit}
\alias{cluster_sequences_cdhit}
\title{calls cdhit}
\usage{
cluster_sequences_cdhit(fa_fil, cl_fil, cutoff = cutoff, cdhdir)
}
\value{
cdhit output
}
\description{
calls cdhit
}
\examples{
\dontrun{
ld_files <- dart2localdiff(gms, etc)
}
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
