% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_sequences_decipher.r
\name{cluster_sequences_decipher}
\alias{cluster_sequences_decipher}
\title{performs sequence clustering using package decipher}
\usage{
cluster_sequences_decipher(
  fa_fil,
  cl_fil,
  cutoff = cutoff,
  method = "complete",
  np = NULL
)
}
\value{
cdhit output
}
\description{
performs sequence clustering using package decipher
}
\examples{
\dontrun{
ld_files <- dart2localdiff(gms, etc)
}
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
