% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dart2genalex.r
\name{dart2genalex}
\alias{dart2genalex}
\title{Returns a dart data object in appropriate files for analysis 
with the Peter and Slatkin method of detecting a range expansion}
\usage{
dart2genalex(
  dart_data,
  basedir,
  species,
  dataset,
  add_pop = FALSE,
  pop = NULL,
  numeric_gt = FALSE
)
}
\arguments{
\item{dart_data}{-- a dart data object [Required]}

\item{basedir}{-- name of the base directory for R&R}

\item{species}{-- species name}

\item{dataset}{-- dataset name}

\item{pop}{-- a vector containing population assignments for each sample}
}
\value{
a list containing names of infiles for range expansion analysis
}
\description{
Input is a dart data object with altcount genotype encoding
(0=hom ref, 1=het, 2=hom alt).
}
\examples{
\dontrun{
dart_gi <- dart2genalex(dart_data, meta.csv)
}
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
