% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.altcount2mincount.r
\name{encode.altcount2mincount}
\alias{encode.altcount2mincount}
\title{Recodes genotypes supplied from count of alternate allele to count of minor allele
Also changes the status from the $encoding element of data list}
\usage{
encode.altcount2mincount(dart_data)
}
\arguments{
\item{dart_data}{-- dart data list  [required]}
}
\description{
Recodes genotypes supplied from count of alternate allele to count of minor allele
Also changes the status from the $encoding element of data list
}
\examples{
encode.dart2altcount(dart_data)
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
