% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loocv_gdm.r
\name{loocv_gdm}
\alias{loocv_gdm}
\title{Perform leave-n-out cross validation for gdm}
\usage{
loocv_gdm(gdm_dir, krig_lambda = NULL, Q = FALSE)
}
\arguments{
\item{gdm_dir}{-- the directory containing the gdm}

\item{gdm_model}{-- the gdm model to use in predictions}

\item{n}{-- number left out, n=1 or n=2}

\item{do_all}{-- if n=2, do all pairs?}

\item{draws}{-- if n=2, and do all is false, nominate a number of draws}

\item{rekrig}{-- if TRUE, rekrig Q martix}
}
\value{
file name
}
\description{
Perform leave-n-out cross validation for gdm
}
\examples{
\dontrun{
cv_gdm_dir <- cv_gdm(gdm_dir, gdm_model, )
}
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
