% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_prediction_domain_file.r
\name{make_prediction_domain_file}
\alias{make_prediction_domain_file}
\title{Clip a raster to a convex hull surrounding a set of points}
\usage{
make_prediction_domain_file(
  species,
  out_dir,
  location_file = NULL,
  gdm_dir = NULL,
  hull_res = 150,
  convexity = 0.85,
  buff_width = 0.1
)
}
\arguments{
\item{species}{-- species name}

\item{location_file}{-- file containing longs and lats}

\item{gdm_dir}{-- directory used to fit GDM}

\item{hull_res}{-- resolution of convex hull}

\item{convexity}{-- convexity of hull around sampled points}
}
\value{
file name
}
\description{
Clip a raster to a convex hull surrounding a set of points
}
\examples{
\dontrun{
new_raster <- make_prediction_domain_file(old_raster, location_file="locations.txt")
}
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
