% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_gt_meta_data.r
\name{merge_gt_meta_data}
\alias{merge_gt_meta_data}
\title{Merges dart and meta data objects into a single object}
\usage{
merge_gt_meta_data(dart_data, meta_data)
}
\arguments{
\item{dart_data}{-- a dart data object [Required]}

\item{meta_data}{-- a meta data object [Required]}
}
\value{
a dart_meta_data object
}
\description{
Input is a dart data object with altcount genotype encoding
(0=hom ref, 1=het, 2=hom alt).
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
