% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_cdhit_file.r
\name{parse_cdhit_file}
\alias{parse_cdhit_file}
\title{parses cdhit output in R}
\usage{
parse_cdhit_file(cdh_out)
}
\value{
cdhit output
}
\description{
parses cdhit output in R
}
\examples{
\dontrun{
ld_files <- dart2localdiff(gms, etc)
}
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
