% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preliminary_PCA.r
\name{preliminary_PCA}
\alias{preliminary_PCA}
\title{Write a dart data list object to an R workspace file}
\usage{
preliminary_PCA(gl_file, data_list, basedir, species, dataset, treatment)
}
\arguments{
\item{gl_file}{-- location of an R file containing genlight object  [required]}

\item{basedir}{-- Base dirctory  [required]}

\item{species}{-- species name  [required]}

\item{dataset}{-- dataset identifier  [required]}

\item{treatment}{-- treatment  [required]}
}
\description{
Write a dart data list object to an R workspace file
}
\examples{
preliminary_PCA(gl_file, basedir, species, dataset, treatment)
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
