% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preliminary_SNMF.r
\name{preliminary_SNMF}
\alias{preliminary_SNMF}
\title{Perform preliminary sNMF analysis}
\usage{
preliminary_SNMF(
  dart_data,
  basedir,
  species,
  dataset,
  treatment,
  pop = NULL,
  Kvals = 8,
  Rvals = 4
)
}
\arguments{
\item{basedir}{-- name of the base directory for R&R}

\item{species}{-- species name}

\item{dataset}{-- dataset name}

\item{treatment}{}

\item{pop}{-- a vector of population assignments}

\item{Ksel}{-- the selected number of ancestral populations}
}
\value{
file name
}
\description{
Perform preliminary sNMF analysis
}
\examples{
\dontrun{
gdm_files <- SNMF2gdm(basedir, species, dataset, treatment, process=TRUE, Ksel=2)
}
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
