% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_RR_data_base.r
\name{query_RR_data_base}
\alias{query_RR_data_base}
\title{Queries DArT data against a sample database, and generates metadata files}
\usage{
query_RR_data_base(
  dart_data,
  basedir,
  species,
  dataset,
  RR_db_file,
  overwrite_meta = TRUE
)
}
\arguments{
\item{datafile}{-- name of xls file containing the meta data [required]}
}
\value{
-- a dart meta data object, consisting of sample names
}
\description{
Queries DArT data against a sample database, and generates metadata files
}
\author{
Jason Bragg (jasongbragg@gmail.com)
meta_data_file <- query_RR_data_base(dart_data, basedir, species, dataset, RR_db_file, overwrite_meta=TRUE)
}
