% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dart_onerow_csv.r
\name{read_dart_onerow_csv}
\alias{read_dart_onerow_csv}
\title{Import DArT (onerow) genotype data and stores them in a list}
\usage{
read_dart_onerow_csv(
  basedir,
  species,
  dataset,
  misschar = "-",
  altcount = TRUE,
  seq2fa = FALSE,
  fnum = 2
)
}
\arguments{
\item{basedir}{-- base directory for analyses [required]}

\item{species}{-- a species name (name of directory within base directory) [required]}

\item{dataset}{-- dataset name assigned by DArT (appears in csv file name) [required]}

\item{misschar}{-- missing data character [default "-"]}

\item{altcount}{-- if TRUE, change genotype encoding to the count of 'alternate' allele  [default TRUE]}

\item{seq2fa}{-- if TRUE, write allele sequences to a fasta file}
}
\value{
A DArT data list, genotype matrix, locus names (by ID), sample names, Reproducibility scores (RepAvg)
}
\description{
It reads the genotype calls in dart format (encoding=dart) and
can covert these to counts of the alternate allele (encoding=altcount)
}
\examples{
\dontrun{
dart_data            <- read_dart_onerow_csv(RandRbase,species,dataset)
} 
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
