% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dart_xls2csv.r
\name{read_dart_xls2csv}
\alias{read_dart_xls2csv}
\title{Import DArT (onerow) xls file and convert to csv}
\usage{
read_dart_xls2csv(basedir, species, dataset, sheet = 4, fnum = 2)
}
\arguments{
\item{basedir}{-- base directory for analyses}

\item{species}{-- a species name (name of directory within base directory)}

\item{dataset}{-- dataset name assigned by DArT (appears in csv file name)}

\item{sheet}{-- sheet number for singlerow dataset}

\item{fnum}{-- number to append at end of filename}
}
\value{
Name of a csv file, containing DArT genotype data that arrived in xls format.
}
\description{
Import DArT (onerow) xls file and convert to csv
}
\examples{
\dontrun{
dart_data            <- read_dart_onerow_csv(RandRbase,species,dataset)
} 
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
