% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_provided_meta_data_file.r
\name{read_provided_meta_data_file}
\alias{read_provided_meta_data_file}
\title{Import metadata for a DArT dataset}
\usage{
read_provided_meta_data_file(
  dart_data,
  basedir,
  species,
  dataset,
  metafile,
  fields = 1,
  nas = "-"
)
}
\arguments{
\item{nas}{-- missing data character [default "-"]}

\item{datafile}{-- name of xls file containing the meta data [required]}
}
\value{
-- a dart meta data object, consisting of sample names
}
\description{
read.meta.data() opens a meta data file associated with a dart dataset
}
\examples{
meta_data <- read.dart.meta(datafile=, nas="-")
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
