% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_fixed_snps.r
\name{remove_fixed_snps}
\alias{remove_fixed_snps}
\title{Remove fixed SNP sites from a gentype list object}
\usage{
remove_fixed_snps(dart_data)
}
\arguments{
\item{genotype}{list object [required]}
}
\description{
Remove fixed SNP sites from a gentype list object
}
\examples{
remove_fixed_snps(dart_data)
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
