% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR_species_params.r
\name{RR_species_params}
\alias{RR_species_params}
\title{Read from a table of RR genotype datasets, get species info}
\usage{
RR_species_params(paramdir, species_file, target_species = "all")
}
\arguments{
\item{paramdir}{-- name of directory containing a file with a standard parameter set [required]}

\item{species_file}{-- name of the file containing information about the genotype data [required]}
}
\value{
-- A list of parameters, which can be used in various data filtering functions
}
\description{
Read from a table of RR genotype datasets, get species info
}
\examples{
\dontrun{
p <- RR_species_params(RandRparam, species_file)
} 
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
