% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_one_snp_per_cluster_random.r
\name{sample_one_snp_per_cluster_random}
\alias{sample_one_snp_per_cluster_random}
\title{Sample one SNP per cluster at random}
\usage{
sample_one_snp_per_cluster_random(dart_data, cluster_fil, seed = 12345)
}
\arguments{
\item{dart_data}{-- dart data list  [required]}

\item{cluster_fil}{-- name of cluster file [required]}

\item{seed}{-- random seed}
}
\description{
Sample one SNP per cluster at random
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
