% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_allele_seq_fasta.r
\name{write_allele_seq_fasta}
\alias{write_allele_seq_fasta}
\title{Write dart locus sequences to a fasta file}
\usage{
write_allele_seq_fasta(x, basedir, species, dataset, clip = 5)
}
\arguments{
\item{x}{-- content of dart data file  [required]}

\item{basedir}{-- name of base directory [required]}

\item{species}{-- name of the species in GenuSpec format [required]}

\item{name}{-- arbitrary name for the dataset [required]}
}
\description{
Write dart locus sequences to a fasta file
}
\examples{
write.dart.data(dart_data, RandRbase, "FicuRubi", "DFi16-2154")
}
\author{
Jason Bragg (jasongbragg@gmail.com)
}
